Fix some errors when shutdown the application
Fix some errors when shutdown the application. The error occurs in
freeing VulkanMemoryAllocation. It is because the pages in VulkanMemoryHeap
were reallocated when a new page created. So I fix it to store a pointer of
the page instead.
Some other errors are occured even applying this commit. It is because of
Resource system, so I will fix it with redesigning the resource system later.

- Fix an error when freeing VulkanMemoryAllocation
- Execute and flush the task queue in rendering thread when destroyed
- Unregister the importers in ResourceManager when destroyed
- Destroy the image view in SwapChainVk when destroyed